task_management_api/requirements.txt
asgiref==3.9.1
distlib==0.3.9
Django==5.2.4
django-filter==25.1
django-s3-file-field==1.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
filelock==3.18.0
platformdirs==4.3.8
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.31.2
virtualenvwrapper-win==1.2.7
Django>=5.0
djangorestframework
djangorestframework-simplejwt
django-filter
